Add files via upload
I have created a Number Guessing Game project that allows the user to guess the number created by the computer. User can get 10 chances to guess the number if user guess the correct number then based on the total attempt marks will be displayed and if failed then user can re-attempt the game.